package/petitboot: new package
Although this is a "boot loader", it really is a tool that should be
part of an image that acts as the bootloader. Thus, it is not in the
bootloader menu.

A UI is required and can be provided by ncurses, X11 or fbdev. Only
ncurses is supported for now. The other two use libtwin which isn't part
of Buildroot for now.

Adapted from https://github.com/glevand/petitboot--buildroot
Updated to work with m68k architecture rather than just PlayStation3
(focused on qemu-system-m68k virt machine)

Signed-off-by: Laurent Vivier <laurent@vivier.eu>
Reviewed-by: Thomas Huth <huth@tuxfamily.org>
[Arnout:
 - Add missing arch and toolchain dependencies.
 - Remove nonexistent BR2_PACKAGE_NCURSES_TARGET_* selects.
 - Swap kexec/kexec-lite priority.
 - Add hash for license file.
 - Use version without v for VERSION.
 - Remove unnecessary --localstatedir=/var.
 - Add --disable-busybox.
 - Remove --enable-debug.
 - Remove --with-ncursesw and explicit libraries.
 - Add explicit arguments for all configure options.
 - Always select busybox tftp.
 - Add optional dependency on dtc.
 - Move dtc-specific options under a single condition.
 - Minor cleanups (whitespace, comments, ...).
]
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>