package/shadow: fix libsemanage dependency
Fix the following build failure raised since the addition of the package
in commit f78c5cb5cae93a9e63dad4361d78e1787759382f:

Makefile:576: *** libsemanage is in the dependency chain of shadow that has added it to its _DEPENDENCIES variable without selecting it or depending on it from Config.in.  Stop.

Fixes:
 - http://autobuild.buildroot.org/results/e528e07874815d7f1cb951084ec985d76fe22d2e

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>