package/rockchip-rkbin: new package
This patch adds a package for the Rockchip ATF binary blobs. These
binaries are needed to build U-Boot for some Rockchip SoCs (e.g.,
RK3588).

As there as so many binary blobs, for each suported SoCs, it is not
practical to introduce a choice, so one must manually define which
blobs (for bl31, tpl and optee) to use from the repository, that
match their actual board.

Signed-off-by: Kilian Zinnecker <kilian.zinnecker@mail.de>
[yann.morin.1998@free.fr:
  - drop custom version
  - simplify copying blobs
  - slightly tweak help texts
  - add hash for license file
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>